nitro-enclaves/requirements.txt
aws-cdk.aws-certificatemanager==1.101.0
aws-cdk.aws-ec2==1.101.0
aws-cdk.aws-iam==1.101.0
aws-cdk.aws-route53==1.101.0
aws-cdk.aws-s3-assets=1.101.0
aws-cdk.core==1.101.0
